Network Intrusion Detection using Decision Trees and Random Forest
The dataset to be audited was provided which consists of a wide variety of intrusions simulated in a military network environment. Each connection is labeled as either normal or as an attack with exactly one specific attack type. Each connection record consists of about 100 bytes.
For each TCP/IP connection, 41 quantitative and qualitative features are obtained from normal and attack data (3 qualitative and 38 quantitative features). The class variable has two categories:
•	Normal 
•	Anomalous